Fix CI flakiness in SchedulableFlowsIT
- Fix WireMock to explicitly bind to 127.0.0.1 instead of localhost to avoid IPv4/IPv6 resolution issues in CI
- Fix database cleanup order in tests to respect foreign key constraints (delete children before parents)
- Fix FlowPlanner context execution to use workflow instance ID
- Add WireMock verbose logging and improve stub matching
- Configure database-specific test profiles for CI
- Fix race conditions with shared event arrays
- Add proper Hibernate/Flyway SQL naming configuration
- Disable Keycloak DevServices to prevent port conflicts

+++ langchain4j/integration-tests/src/main/resources/db/migration/V1__create_tables.sql
@@ -36,7 +36,7 @@ CREATE TABLE task_info_entity
     iteration           INTEGER      NOT NULL,
     task_type           INTEGER      NOT NULL CHECK (task_type IN (1, 2)),
     is_end_node         BOOLEAN,
-    retry_attempt       INTEGER,
+    retry_attempt       SMALLINT,
     instant             TIMESTAMP(6) WITH TIME ZONE,
     next_position       VARCHAR(255),
     context             VARBINARY,
